start ai integration
- added full ai service: system prompt, user prompt building system,
saved history, response parsing into Go struct, etc.
- made the symbols list accessible globally
- made the agent/llm.go file for later use when i build the actual agent
service

still a lot of todo's for the RNG agent to be completed but I wanted to
get the AI service out of the way. next, i will finish all the todo's
for RNG and then start working on the LLM agent impl.